Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main()
 {
   int a,b,c;
   scanf("%d %d %d",&a &b &c);
-  int d=(+b+c)/3;
-  printf(" Average: %d",d)
+  float d=(+b+c)/3;
+  printf(" Average: %.2f",d)
    return 0;
 }
